requirements.txt
openai>=1.30.0
tiktoken>=0.7.0
pydantic>=2.5.3
python-dotenv>=1.0.1

# LangChain core and integrations
langchain>=0.2.7
langchain-openai>=0.1.13
langchain-community>=0.0.21

# Vector store
faiss-cpu>=1.8.0

# UI
gradio>=4.44.0

unstructured>=0.10.0
pypdf>=3.12.0